Noi limits logic

+++ contracts/base/vex_account/VexesAccountStorage.sol
@@ -20,6 +20,6 @@ abstract contract VexesAccountStorage is IVexesAccount {
     uint128 constant CONTRACT_MIN_BALANCE = 1 ever;
     uint128 constant SCALING_FACTOR = 10**18;
     uint64 constant HUNDRED_PERCENT = 1_000_000_000_000; // 100%, this allows precision up to 0.0000000001%
-    uint32 constant LEVERAGE_BASE = 100;
+    uint8 constant LEVERAGE_BASE = 100;
     uint32 constant HOUR = 3600;
 }
